requirements.txt
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
colour==0.1.5
docutils==0.20.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
gevent==23.7.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.2
MarkupSafe==2.1.3
matrx==2.3.2
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
pycparser==2.21
Pygments==2.16.1
python-dateutil==2.8.2
python-engineio==4.6.1
python-socketio==5.8.0
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
